requirements.txt
python-telegram-bot==13.5
Django
djangorestframework
requests